# Flush performance