# Summary

Moving cache implementation structure from method -> {addressable, params} -> task
to method -> addressable -> params -> task.

This facilitates updating the flush implementation so it doesn't have to walk the entire addressable structure to flush the cache for an actor.

# Benchmarks

Existing flush implementation throughput

```
Benchmark                                              Mode  Cnt     Score     Error  Units
CacheResponseFlushBenchmark._0_flushThroughput10      thrpt   20  2246.177 ± 105.392  ops/s
CacheResponseFlushBenchmark._1_flushThroughput100     thrpt   20  1987.013 ±  96.470  ops/s
CacheResponseFlushBenchmark._2_flushThroughput500     thrpt   20   971.735 ±  22.070  ops/s
CacheResponseFlushBenchmark._3_flushThroughput1000    thrpt   20   386.572 ±  36.502  ops/s
CacheResponseFlushBenchmark._4_flushThroughput10000   thrpt   20    36.829 ±   1.787  ops/s
CacheResponseFlushBenchmark._5_flushThroughput100000  thrpt   20     3.088 ±   0.023  ops/s
```

New flush implementation throughput

```
CacheResponseFlushBenchmark._0_flushThroughput10      thrpt   20  2807.531 ± 171.453  ops/s
CacheResponseFlushBenchmark._1_flushThroughput100     thrpt   20  2862.684 ± 185.700  ops/s
CacheResponseFlushBenchmark._2_flushThroughput500     thrpt   20  2906.574 ± 135.972  ops/s
CacheResponseFlushBenchmark._3_flushThroughput1000    thrpt   20  2806.602 ± 162.794  ops/s
CacheResponseFlushBenchmark._4_flushThroughput10000   thrpt   20  2709.984 ± 144.680  ops/s
CacheResponseFlushBenchmark._5_flushThroughput100000  thrpt   20  2348.267 ± 377.029  ops/s
```

Old implementation flush throughput degrades with more entries in the cache. 
New implementation flush throughput stays much more consistent with more entries.

Old implementation get/put throughput

```
Benchmark                          Mode  Cnt       Score      Error  Units
CacheResponseBenchmark.get        thrpt   20     714.949 ±   37.018  ops/s
CacheResponseBenchmark.getCached  thrpt   20  191107.633 ± 3111.927  ops/s
CacheResponseBenchmark.put        thrpt   20   39724.746 ± 2623.486  ops/s
CacheResponseBenchmark.putCached  thrpt   20   28512.045 ± 5513.049  ops/s
```

New implementation get/put throughput

```
CacheResponseBenchmark.get        thrpt   20     688.818 ±    8.487  ops/s
CacheResponseBenchmark.getCached  thrpt   20  192996.263 ± 7377.763  ops/s
CacheResponseBenchmark.put        thrpt   20   40627.001 ± 2837.266  ops/s
CacheResponseBenchmark.putCached  thrpt   20   30988.189 ± 3545.611  ops/s
```

Throughtput for get/put not affected by the improvement to flush.

# Load Testing

Existing implementation can't keep up with the flush rate.

<img width="689" alt="screen shot 2018-04-24 at 8 52 07 am" src="https://user-images.githubusercontent.com/318994/39188015-cd7cb4e2-479c-11e8-9fe7-957be8fc1663.png">

<img width="685" alt="screen shot 2018-04-24 at 8 50 35 am" src="https://user-images.githubusercontent.com/318994/39187980-b5c19642-479c-11e8-9bb2-82dc00dbc318.png">

Existing implementation performs much better.

<img width="686" alt="screen shot 2018-04-24 at 8 52 41 am" src="https://user-images.githubusercontent.com/318994/39188052-df60e1d8-479c-11e8-835f-e55d25c0d74b.png">

<img width="685" alt="screen shot 2018-04-24 at 8 50 53 am" src="https://user-images.githubusercontent.com/318994/39188033-d64d336c-479c-11e8-9c49-f75e98752112.png">