@JordanAngold 2018-04-24
Change looks great, numbers look great. I'd like to see benchmark numbers for 500k, 1M (those will be closer to our actual load values, right?).